requirements.txt
selenium>=4.38.0
requests>=2.31.0
discord.py>=2.3.0
lxml>=4.9.0


